rename: ReviewHashtag 폴더 수정
기존에는 분리를 하였지만 하나로 합침
