+++ backend/src/main/java/com/multi/mariage/review/domain/Review.java
@@ -5,7 +5,6 @@ import com.multi.mariage.like.domain.Like;
 import com.multi.mariage.member.domain.Member;
 import com.multi.mariage.product.domain.Product;
 import com.multi.mariage.review.dto.request.ReviewSaveRequest;
-import com.multi.mariage.review_hashtag.domain.ReviewHashtag;
 import com.multi.mariage.storage.domain.Image;
 import com.multi.mariage.weather.domain.Weather;
 import jakarta.persistence.*;
