+++ backend/src/main/java/com/multi/mariage/review/domain/ReviewHashtag.java
@@ -1,4 +1,4 @@
-package com.multi.mariage.review_hashtag.domain;
+package com.multi.mariage.review.domain;
 
 import com.multi.mariage.hashtag.domain.Hashtag;
 import com.multi.mariage.review.domain.Review;
