+++ backend/src/main/java/com/multi/mariage/review/domain/ReviewHashtagRepository.java
@@ -1,4 +1,4 @@
-package com.multi.mariage.review_hashtag.domain;
+package com.multi.mariage.review.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
