+++ backend/src/main/java/com/multi/mariage/review/service/ReviewHashtagService.java
@@ -1,10 +1,10 @@
-package com.multi.mariage.review_hashtag.service;
+package com.multi.mariage.review.service;
 
 import com.multi.mariage.hashtag.domain.Hashtag;
 import com.multi.mariage.hashtag.service.HashtagService;
 import com.multi.mariage.review.domain.Review;
-import com.multi.mariage.review_hashtag.domain.ReviewHashtag;
-import com.multi.mariage.review_hashtag.domain.ReviewHashtagRepository;
+import com.multi.mariage.review.domain.ReviewHashtag;
+import com.multi.mariage.review.domain.ReviewHashtagRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
