+++ backend/src/main/java/com/multi/mariage/review/service/ReviewModifyService.java
@@ -10,7 +10,6 @@ import com.multi.mariage.review.domain.Review;
 import com.multi.mariage.review.domain.ReviewRepository;
 import com.multi.mariage.review.dto.request.ReviewSaveRequest;
 import com.multi.mariage.review.dto.response.ReviewSaveResponse;
-import com.multi.mariage.review_hashtag.service.ReviewHashtagService;
 import com.multi.mariage.storage.domain.Image;
 import com.multi.mariage.storage.service.ImageService;
 import com.multi.mariage.weather.domain.Weather;
