+++ backend/src/test/java/com/multi/mariage/common/annotation/ServiceTest.java
@@ -18,7 +18,7 @@ import com.multi.mariage.product.service.ProductModifyService;
 import com.multi.mariage.review.domain.ReviewRepository;
 import com.multi.mariage.review.dto.response.ReviewSaveResponse;
 import com.multi.mariage.review.service.ReviewModifyService;
-import com.multi.mariage.review_hashtag.service.ReviewHashtagService;
+import com.multi.mariage.review.service.ReviewHashtagService;
 import com.multi.mariage.storage.dto.response.ImageSavedResponse;
 import com.multi.mariage.storage.repository.StorageRepository;
 import com.multi.mariage.storage.service.StorageService;
