+++ backend/src/test/java/com/multi/mariage/review_hashtag/domain/ReviewHashtagRepositoryTest.java
@@ -6,6 +6,7 @@ import com.multi.mariage.hashtag.domain.Hashtag;
 import com.multi.mariage.member.domain.Member;
 import com.multi.mariage.product.domain.Product;
 import com.multi.mariage.review.domain.Review;
+import com.multi.mariage.review.domain.ReviewHashtag;
 import com.multi.mariage.storage.domain.Image;
 import com.multi.mariage.weather.domain.Weather;
 import org.junit.jupiter.api.BeforeEach;
